ref: Reorganize organization dashboard files
